feat: implement schedule plan, update plan | enhance rabbitmq integration | add worker microservice | add basic frontend

+++ src/PlanIt.Backend/PlanIt.Plan/PlanIt.Plan.Application/RabbitMq/PublishHelper.cs
@@ -2,38 +2,29 @@
 using Microsoft.EntityFrameworkCore;
 using PlanIt.Plan.Application.Configurations;
 using PlanIt.Plan.Application.Interfaces;
+using PlanIt.Plan.Domain.Enums;
 
 namespace PlanIt.RabbitMq;
 
 public class PublishHelper : IPublishHelper
 {
     private readonly IBus _bus;
-    private readonly RabbitMqQueuesConfiguration _queues;
     private readonly IApplicationDbContext _dbContext;
 
     public PublishHelper(IBus bus, RabbitMqQueuesConfiguration queues, IApplicationDbContext dbContext)
     {
         _bus = bus;
-        _queues = queues;
         _dbContext = dbContext;
     }
 
-    public async Task PublishInstantPlanTriggered(InstantPlanTriggered message, CancellationToken cancellationToken)
-    {
-        await _bus.Send(message, cancellationToken);
-    }
-
-    public async Task PublishOneOffPlanTriggered(OneOffPlanTriggered message, CancellationToken cancellationToken)
+    public async Task PublishScheduledPlanTriggered(ScheduledPlanTriggered message, CancellationToken cancellationToken)
     {
         await _bus.Send(message, cancellationToken);
 
-        await _dbContext.OneOffPlans
-            .Where(oneOffPlan => oneOffPlan.Id == message.OneOffPlanId)
-            .ExecuteDeleteAsync(cancellationToken);
-    }
-
-    public async Task PublishRecurringPlanTriggered(RecurringPlanTriggered message, CancellationToken cancellationToken)
-    {
-        await _bus.Send(message, cancellationToken);
+        //TODO: make another consumer to delete OneOff-Type Scheduled Plans when it executes on Worker Side
+        if (message.ScheduledPlanType == ScheduledPlanType.OneOff)
+            await _dbContext.ScheduledPlans
+                .Where(sp => sp.Id == message.ScheduledPlanId)
+                .ExecuteDeleteAsync(cancellationToken);
     }
 }
